backend/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
Flask==2.3.3
Flask-PyMongo==2.3.0
flask-restx==1.2.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
idna==3.4
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jwt==1.3.1
MarkupSafe==2.1.3
passlib==1.7.4
pycparser==2.21
pymongo==4.5.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==2.3.7
